+++ eslint.config.mjs
@@ -3,7 +3,12 @@ import withNuxt from './.nuxt/eslint.config.mjs'
 
 export default withNuxt(
   // ...Custom flat configs append after nuxt's configs
- 
+   {
+    files: ['**/*.ts', '**/*.tsx'],
+    rules: {
+        //'no-console': 'off' // allow console.log in TypeScript files
+     }
+   }
 ) .prepend(
   // ...Prepend some flat configs in front
 )